Algorithm is beginning to come together.  No connection to php yet
At least checks if a server is full and adds transactions to it if not.
If it IS full, it will try to add the transactions to the next server in
the same region.  If all the servers in the region are full (as in, at
180, which means we lose no transactions), then we will fill up the
servers as much as possible, and then calculate the left over
transactions, and figure out what is profitable to do from there.